Remove contributors.scss in favor of a utility class
The contributors.scss stylesheet only applied some top padding to the
contributors loader. The padding was computed based on the header's
height, which was deemed unnecessary. To simplify things a bit, we are
now applying the padding with a GitLab UI utility class, which lets us
remove the stylesheet altogether.