Load all license details from Gitaly
Gitaly has changed it's interface so it provides all the details about
the software license of a repo.

This change, behind a feature flag, removes the use of the Licensee gem
for this purpose, and solely depends on Gitaly to provide all
information on the license.

Licensee is still used for license templates and linking to a license in
for example the specs.

Changelog: performance
Roll-out-Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/374300