Replace visitUrl with Gl-Link in ci-badge-link
A previous change, which replaced the badge's `href` with a JS solution,
broke native browser behavior like opening the link in a new tab when
certain keys are pressed, etc.

This MR replaced the link with an instance of `GlLink` and
re-introduces the `href` to get back the native browser functionality.

Changelog: fixed

Changelog: fixed