Replace attr_encrypted with vendored fork
This gem is unmaintained and had no working build setup.
We also identified issues with Ruby 3 support.

We therefore forked it into gitlab-org, applied
several patches, and eventually are vendoring
it into gitlab-rails with this commit.

This also allows us to remove various monkey patches
since these issues are now fixed in the forked gem.