Adjust old SCSS to not break padding of gl-alerts
This old CSS was built for alerts without icons. It is still required as
long as we use `createFlash` in flash.js, which is deprecated but still
in use in many places.
As soon as we will have switched all `createFlash` to `createAlert` we
could drop this SCSS. For now, we have to work around it.