Refetch deployments on user interaction
Polling for the deployments block is configured for 4 minutes at the
longest. While there is work to be done there refining the rate, we can
assume that if the user triggers an action, we should refetch the
deployments immediately after the request finishes.

This is done via the event hub, as otherwise we need to send a Vue event
up too many layers of components.

Changelog: changed