Add batch loading to CI/CD variables
In the context of the CI/CD refactor, we did not
take into account that graphQL queries limit is 100
ms. Therefore, this commit adds batch loading
to the list of variables for project and group. Instance
variables have a hard limit at 25, but can be changed
so there is also batch loading there.