Fix 500 errors when sorting by merged date with approvers
When adding a `Approved By` or `Approver` filter and sorting by merged
date, a 500 error will occur because of a SQL error:

```
PG::GroupingError: ERROR: column "merge_request_metrics.id" must
appear in the GROUP BY clause or be used in an aggregate function

```

This happens because the approval query uses a `GROUP BY` query find
merge requests that match all approvers. However, this `GROUP BY`
query breaks the SQL query since it's expected that
`merge_request_metrics.id` is present in that clause.

This is similar to a bug fixed with sorting by milestones:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/97726.

The UI also sorts by `merged_at`, so we need to account for that too.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/375192

Changelog: fixed