Remove Git cheat sheet PDF LFS file
`spec/fixtures/git-cheat-sheet.pdf` was checked in as an LFS file in
2019, but it's not really essential since it's just a test
fixture. Having this file causes some unintended headaches:

1. We now have to pay for LFS bandwidth for GitLab mirrors.

2. When a conflicting GitLab.com SSL cert is installed (e.g. due to
JAMF), a `git clone` will fail because of HTTPS verification. We've
seen a number of people blocked as a result.

While it's useful to dogfood LFS, the downsides seem to outweight the
upsides.