Fix 500 errors when sorting by closed date with approvers
When adding a `Approved By` or `Approver` filter and sorting by merged
date, a 500 error will occur because of a SQL error:

```
PG::GroupingError: ERROR: column "merge_request_metrics.id" must
appear in the GROUP BY clause or be used in an aggregate function

```

This happens because the `order_by_metric` uses the ID as a secondary
sort order.

There is also no `merge_request_metrics.closed_at`; use
`merge_request_metrics.latest_closed_at` instead.

This is similar to a bug fixed with sorting by milestones:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/97726.

The UI also sorts by `closed_at`, so we need to account for that too.