Allow users to create annotated tags from release
Users not only sometimes want to annotate tags with messages, but also
wish for them to be different from the officially posted release notes.

A new text area is added to allow users to write out a tag message, and
a collapse is utilized to show and hide the tag creation inputs, only
showing them the user wishes to create a new tag.

Changelog: added