Replace custom CVSS validation with cvss_suite gem
The first iteration of CVSS validation was a custom implemention.
However, there exists a gem which does everything we want to do and
should involve lower maintenance overhead. This change switches to the
gem and removes the custom implementation. Since cvss_suite works with
versions 2, 3.0, and 3.1, this change adds tests for cvss_v2 validation
and refactors the validation to use a validator.