Upgrade Sourcegraph integration to 0.0.82
This change upgrades the Sourcegraph dependency to the latest version.

We (Sourcegraph) have recently worked on [changes to our extensions
architecutre](https://docs.sourcegraph.com/extensions/deprecation) which
also affects all of our external extensions (e.g. IDE, browser, and
native extensions).

The native GitLab extension is special though, since, as opposed to the
other extensions, we do not control the deployment of this client and
instead rely on GitLab's rollout cycle.

For our recent release, [we made sure to not break existing GitLab
instances](https://github.com/sourcegraph/sourcegraph/pull/41572) that
are still using the old extension client. We can maintain these changes
for a while but ultimately we want to delete the old extension code.

To prepare our customers for this change and to have as little impact on
GitLab on-premise instances as possible, we want to upgrade our GitLab
native extension sooner so that when we come to the actual deprecation
in the future, our GitLab customers have had plenty of opportunities to
upgrade.

The main difference and why we want to upgrade to 0.0.82 as soon as
possible is that we now do a feature test against the Sourcegraph
instance to find out if the migrated extensions APIs are used. If this
is the case, we _no longer download extension code from Sourcegraph
servers_ but instead fall back to inlined extension code that are
included in the updated npm package.

Since the native integration version that was bundled with GitLab was
already quite dated, this change also includes several [bug
fixes](https://github.com/sourcegraph/sourcegraph/blob/main/client/browser/CHANGELOG.md)
that we have worked on over the last months. These are tested against
newer GitLab versions regularly as it shares the code with our [browser
extension](https://docs.sourcegraph.com/integration/browser_extension).

It took me quite a while to get it set up but I verified my changes
locally by setting up a GitLab and Sourcegraph development environment
and manully testing through various scenarios. You can learn more about
it [in this Sourcegraph
commit](https://github.com/sourcegraph/sourcegraph/pull/42106).

Changelog: changed