Show copy button for unsafe external URLs
We've recently loosened the verifiction of external URLs of environments
to allow schemes other than http(s). When these links are used,
navigating to them with a browser is not a good idea, but users might
still need to copy them and paste them elsewhere, like for a database
URL.

And so instead, when the link is unsafe, render a copy button.

Changelog: changed