Trigger mergeRequestMergeStatusUpdated when merge_status changes
When `merge_status` transitions, we need to let the subscribers
know about it. To do this, we trigger the subscription on
specific transitions.

This is behind `trigger_mr_subscription_on_merge_status_change`
because the `merge_status` transitions is critical to mergeability.
In case, trigger cause some unforeseen issue, we can toggle it.