Decompose timebox concern and remove project iteration
Project-level iterations were implemented but never rolled out.
In this commit, the implementation is removed.

The code pertaining to group or project is also decomposed
and moved into the respective models Milestone and Iteration
out of Timebox concern. Originally milestones and iterations were
envisioned to share most traits as included in the concern
but have diverged since.