Fix crash if createAlert dismissed multiple times
If an alert, generated by `createAlert`, was dismissed multiple times it
throw an error. This fixes this problem.

This situation raises when the user manually dismiss the alert, but
there is a programmatic dismiss that clears the alert. One could track
this situation with `onDismiss` but I think it is a little bit
cumbersome. I would rather made the dismiss method more robust.