Add terms disclaimer to Group SAML sign in page
With recent unification of sign-in flow with external providers,
terms are accepted automatically as part of the sign-up process.
The global sign-in page already included a disclaimer that registering
or signing in signifies acceptance of terms, but Group SAML sign-in
pages did not. This brings the same disclaimer to Group SAML.

Changelog: fixed
EE: true