Allow pipeline schedule owner to call take_ownership API
Currently, a user is granted the take_ownership_pipeline_schedule
ability if they have the admin_pipeline_schedule ability and are not
already the owner of the schedule. This results in the take_ownership
API returning 403 if the user calling the API is already the owner
of the schedule.

This commit deprecates the `take_ownership_pipeline_schedule` ability
and instead allows users with the `admin_pipeline_schedule` ability to
take ownership of the schedule, regardless of whether the user is
already the owner of the schedule.

The ability has been deprecated instead of being removed since the
GraphQL type Types::PermissionTypes::Ci::PipelineSchedules has a field
for this ability. The GraphQL field has been deprecated and can be
removed in 17.0. Once the GraphQL field is removed, we can remove the
`take_ownership_pipeline_schedule` ability from
Ci::PipelineSchedulePolicy.

Changelog: fixed