Enforce active jobs limit earlier in pipeline creation
Remove the feature flag `ci_limit_active_jobs_early` which
moves the check for active jobs earlier in the pipeline
creation chain.

This change allows to persist empty pipelines instead of
full pipelines when it's required to persist a failed
pipeline. When persistence is not required we won't
persist the pipeline.

Changelog: changed