ci: Extract assets 'tmp/cache' to a separate native cache
This is to ensure that the component caching mechanism doesn't store
these files. These files can aren't compiled assets, and can be reused
partially if present so, it's ok to use native cache for those.

This should reduce the size of the assets packages.

This extraction from the main native assets cache is needed in the case
we want to stop using this main native assets cache in the future.

Signed-off-by: Rémy Coutable <remy@rymai.me>