Trigger mergeRequestMergeStatusUpdated when MR marked as draft/ready
When draft status of a merge request changes, it affects the
mergeability of the MR.

Since we need to update the merge widget in real-time, we need to
let the subscribers know about it. To do this, we trigger the
subscription whenever a MR is marked as draft or ready.

Changelog: added