Allow admins to delete runners in bulk
This change adds a checkbox to each runner entry in the Admin -> Runners
page that allows admin users to select multiple runners and delete them
in bulk.

This feature is added by removing the `admin_runners_bulk_delete`
feature flag along and allowing users to access this feature.

Changelog: added