QA: Use an unexisting license in test
The new license detector in Gitaly recognizes the GFDL-1.2-only license.
So when the feature flag 'gitaly_go_find_license' would be flipped, the
results would differ.

Instead of using an existing license, use a completely made-up license
so both the old and new implementation will provide the same results.