git: Nickname the Other license to LICENSE
When the license detector does not recognize the license the users will
see "Other" license in the badge on the project page. This is a bit
confusing if you don't know this applies to the LICENSE.

Instead display LICENSE in the badge, so the purpose is instantly clear.
For the new implementation, this change was also made [1] on Gitaly's
side.

[1]: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4851