Migrate Sidekiq jobs outside of current routing rules
This migration is primarily meant for self-managed users.
As we are changing the default in Sidekiq process to listen to
queues based on routing rules (instead of all queues), we need
to migrate the jobs outside of routing rules mapping to the correct
queue based on the current routing rules.
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1930

Changelog: changed