Render markdown for snippets if links contain hash to anchor
Snippet links containing a hash always rendered the simple viewer.
This commit fixes this by only rendering the simple view if the hash
contains an `#LC` prefix which is used to refer to a certain line
in the source file. Otherwise the rich viewr is rendered (if possible).

Changelog: fixed