Increase the deadline of the FindLicense RPC
This increases the deadline of `FindLicense` to the medium one from the
fast one. With the default settings (like we're running on
GitLab.com), this means an increase from 10s to 30s.

With this, when the cache for this RPC expires, it will be more likely
that a single RPC call succeeds repopulating the cache. Avoiding
incidents like https://gitlab.com/gitlab-com/gl-infra/production/-/issues/7837