fix: Don't fetch environments on Admin variables page
A request to fetch environments would be sent on the Admin variables
page when the `ci_variable_settings_graphql` FF is disabled, but that
doesn't make sense as we're not in a project context.

A system test started to fail recently because of an ambiguous match on
alert boxes.

The fix is to only fetch environments if we're in a project's context.

Possibly related MRs are:

- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89332
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/94267
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/98856

Changelog: fixed
Signed-off-by: Rémy Coutable <remy@rymai.me>