Ensure schemas are up-to-date after migration tests complete
If a CI migration leaves the database behind the current database
schema, we need to ensure that the schema is brought up-to-date after
the migrations run to prevent other tests from starting from an
inconsistent state.

This commit consolidates the database cleanup for multiple databases
by ensuring:

1. After a migration context is complete, either recreate the
databases (if the number of columns approaches the max) or migrate all
the schemas back to the latest and drop all data from both databases.

2. After a migration example runs, delete any leftover data for that
specific database.  If a CI migration leaves the database behind the
current database schema, we need to ensure that the schema is brought
up-to-date after the migrations run to prevent putting other tests in
an inconsistent state.

Related MR:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/100233

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/377796

Changelog: fixed
Signed-off-by: Rémy Coutable <remy@rymai.me>