Change Git examples to use `refs/heads/*`
Custom rulesets with Git passthroughs do not work when the ref is of the
form `refs/remotes/<name>/<branch>`, despite this being a valid ref name
that exists in the repository’s `.git/refs` directory.

Digging into the problem, it appears that when we clone [1] the
repository, `go-git` attempts to computes [1] the refspec of the `ref`
being requested in a way that isn’t compatible with a `ref` that points
to a remote branch.

For example, if we specify `ref = “refs/remotes/origin/main”` in the
ruleset config, go-git will compute the following as the refspec:

  +refs/heads/origin/main:refs/remotes/origin/origin/main

Notice the duplicated “origin/origin” on the remote-side of the mapping.

This issue causes the clone to fail and the analyser to produce a log
message that is a bit of a red-herring:

  couldn't find remote ref "refs/heads/origin/main"

Cloning works as expected if `ref/heads/<tag>` or a commit hash is used.

[1] https://gitlab.com/gitlab-org/security-products/analyzers/ruleset/-/blob/d019b7464c5301fbd3ef13a55231e904e6a719b3/ruleset.go#L773-778
[2] https://github.com/go-git/go-git/blob/da810275bf682d29a530ed819aff175f47bd7634/repository.go#L895