Fix Wiki feature raises exception when titles start with ~
In Wiki#sluggified_title, we are using File#expand_path to remove
relative components of the title. Unfortunately, if the title starts
with "~" character, that method treats the path as a part of user home
dir.

The fix is to switch to File#absolute_path, identical to the troublesome
one without user home dir expansion.

Changelog: fixed