Add compound assertion when delete_refs uses an invalid ref format
In https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4921, Gitaly
will learn to validate ref formats with a backslash \ in it. Currently,
this case falls through to the `git update-ref` `prepare` stage, which
gets sent back as a GitError.

This is not ideal however, since it's an error we should catch when validating
refs.

Once the Gitaly MR is merged, it will start sending a structured error,
which will get prased in ref_service.rb. Until then, we need the spec to
expect both kinds of errors. Once the Gitaly MR is merged, we can remove
the assertion expecting Gitlab::Git::Repository::GitError.