Stop using broken_repo project trait
This trait relies on being able to delete a git managed directory out of
gitaly. This will not be possible in a replicated environment like
praefect where it is impossible to know exactly where the repository
will be stored. Since there is no way to replicate this test through
gitaly RPCs we have to rely on stubbing the RPC instead.