Ensure MembersAddedEvent only published on success
Previously this event was being published every time this service was
called and enqueing jobs from the event store that react to this event.
If no members are actually successfully added, we should not trigger
this event as there shouldn't be any membership changes for the
listeners to react to.

Issue: https://gitlab.com/gitlab-org/customers-gitlab-com/-/issues/4934
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/100647