Remove and add ci_job_variables.raw
In a previous work, we added a migration that re-adding
the ci_job_variables.raw column to make all the `raw` columns to FALSE
because we do not set TRUE anywhere.

During this migration, we missed a place where multi-version deployments
can cause a wrong data.

So in this new migration, we will fix the old data
by re-adding the column. Also, we prevent future wrong data by adding
"raw: true" to the missed place.

Changelog: fixed