Store future subscriptions on instance activation
Prior to this change, when activating an instance and a renewal was
already done for the subscription linked to the activation code, the
information about the renewal was only known on the next subscription
details sync. A banner when the subscription is about to expire was
still displayed before that sync.

Now, when activating in this scenario, the future subscriptions will
also be synced directly and will be displayed in the subscription
history table. A banner will not be displayed anymore due to knowing
about the future subscription.

Changelog: changed
EE: true