Update Jitsu authenticator to production endpoint
Previously, we were hitting the Jitsu endpoint with the "/configurator"
sub-path.

This URL is only available when running the Jitsu docker image that
combines both the server and configurator UI.

In production use cases, the Configurator UI is a standalone service, or
not used at all.

See https://jitsu.com/docs/deployment/deploy-with-docker for more
details.

Changelog: changed
EE: true