Generate RuboCop TODO config even without any offenses
Sometimes we enable new cop rules without any offenses and put this rule
in "grace period". This is done in the TODO config in
`.rubocop_todo/**/*.yml` because it's a temporary configuration which
should be removed short after.

However, when regenerating TODOs for cops without any offenses we
discarded to generate them.

This commit fixes this issue and generates a cop TODO config if:
* Cop was previously disabled due to too many offenses
* Cop is in grace period
* Cop excludes files