Nullify merge request and user associations in batches on user deletion
We have been seeing failures to delete a user because the foreign key
relation with `users.id` and `merge_requests.updated_by_id` is failing
to nullify within the 15-second statement timeout. Just as we did with
nullifying IDs in batches with the issues table in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/84709, we need
do the same with `merge_requests.updated_by_id` and
`merge_requests.assignee_id`. Note that `author_id` and
`merge_user_id` are assigned to the ghost user.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/358760

Changelog: performance