Move groupsPath util to its own file
This moves the groupsPath util to its own utils.js file. This makes it
more reusable and testable. We're doing this in preparation for
migrating the group selector away from select2 as we'll be able to
import this helper in the new Vue component that will coexist with the
existing select2 implementation for a short while.