Use indifferent keys for wal_locations
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/364799

**Problem**

We have a problem with Sidekiq workers when they use stale data for
`sticky` and `delayed` consistency modes.

**Experiment**

The `wal_locations` hash might contain string keys after Sidekiq
transformations. Because of that we can skip location verification and
mark host as up to date while it's not in sync.

I change the location detection logic to accept both symbols and
strings. If that modification resolves the problem, then the assumption
about string keys is correct.