Nullify and delete some user associations in batches
When a user has a lot of associated events, Todos, or other items,
deletion of the user can timeout due to the foreign key cleanup taking
longer than the allowed 15 seconds.

To avoid this, we do what we did in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/100962 and
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/101016 and mark
these associations as `dependent: :nullify` or `dependent: destroy` in
order to ensure `Users::DestroyService` cleans them up in batches.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/358760

Changelog: performance