Reset form button if error encountered
Previous to https://gitlab.com/gitlab-org/gitlab/-/merge_requests/98747,
the `input[type=submit]` element was re-enabled after an unsuccessful
submit.  Otherwise, the user had to retype the name of the token to
re-enable the submit button.

Here, we apply the previous behavior to `button[type=submit]`. In
addition, it prevents an error in the web console.

Changelog: changed