doc/housekeeping: Extend list of housekeeping tasks
The list of high-level housekeeping tasks is not complete. Most
importantly, it is missing information that housekeeping also removes
stale data like lockfiles, which is important information for an admin
in cases where a repository is currently unservable because any such
leftover files.

Amend the list to cover previously-missing tasks.