doc/houseweeping: Improve object pool documentation
While object pools are just another Git repository, GitLab needs to
handle them specially in the context of repository housekeeping. And
while the administrator-facing documentation points that out, it is
overall way too technical.

Rewrite the documentation to point out what is different to normal
repository housekeeping and when exactly object pools are getting
updated.