Fix naming issue with Test Summary widget
The telemetry uses the widget "name" property.

In this case, that's WidgetTestSummary.

However, when I created the legacy telemetry events list,
I used the folder name for this widget, which resolves to
testReport.

This updates the test for this to use the right name, and the
non-standard events list to also reference the widget name.

Changelog: fixed