Geo: Fix redirect to secondary after SAML sign in
Due to changes to the authentication flow for single sign-on via
secondary Geo sites, users who sign in via Geo secondaries with SAML are
landing on the primary site. This is very confusing.

This commit causes the primary site to remember which site the user
attempted to sign in through. And then after successful sign in, the
user is redirected to that site.

This implementation gets around the fact that a typical SAML
configuration will have a single ACS (Assertion Consumer Service) URL
defined in the SAML Identity Provider configuration, as well as in the
primary Geo site's `gitlab.rb` (Omnibus GitLab) or `values.yaml` (GitLab
Helm Chart). This is possible because the secondary site only acts as a
proxy, and the primary knows that the secondary site's URL is safe to
redirect to after sign in. It should work without action by sysadmins,
whereas specifying a particular ACS configuration would require action,
and might not even be possible, depending on the Identity Provider.

Changelog: fixed
EE: true