Display pipeline name on pipeline details and list pages
The pipeline name is displayed instead of the commit title on the
list page.
If a pipeline does not have a name, the commit title is displayed.

On the details page, the pipeline name is displayed as the header
instead of the commit title. The commit title is moved to the row
that currently only displays the commit SHA.