Rename PipelinePresenter#name to event_type_name
Since pipelines can now have a name set by the user, add
PipelinePresenter#event_type_name so that we can remove
PipelinePresenter#name so that it doesn't clash with Pipeline#name.

We continue to keep the name in the PipelineEntity, along with the
newly introduced event_type_name, for 1 milestone to allow for
non-atomic upgrades. For example, if a user gets the new frontend
code but the pipelines API call hits a backend server that has not
yet been upgraded, the UI should still work. We can remove `name`
from PipelineEntity in the next milestone (15.7).