Add logging for file variable expansion
At the Runner side, we want to change the behavior of file variables
expansions. File variables are expanded with their context at the Runner
side for now, we want to expand them with their path. However, we can
not do this change until we know the impact to users. At the GitLab
side, we do not do this expansion, but we can track if a file variable
is referenced in another variable.

Changelog: other