Fix wrong behavior when CI keyword "when" is an array
We have the "inclusion" validation for the "when" keyword but no
validation for its type. So, the attribute also accepts an array as
the value. The "inclusion" validation of "ActiveModel::Validations"
also accepts array values and compares each element.

In this commit, we are adding the "type: String" validation for
the "when" keyword for Job and Bridge entries.

Changelog: fixed