Show mergedYaml content whenever available
In the pipeline editor, we were not showing the
merged yaml if the ci config was invalid, however
sometimes you do get the `mergedYaml` field back
even with an invalid configuration. In these cases,
we should let the user see it. This MR fixes this
and adjust the tests.

Changelog: fixed