Increase lock ttl in LooseForeignKeys::CleanupWorker
The `in_lock` worker is used to guarantee that there are never 2
instances of this worker running at the same time cleaning up the same
data. The use of `ttl` for `in_lock` ensures that whatever lock we
acquire in Redis is eventually expired even if we don't explicitly
expire it ourselves. The `ttl` is not exercised on a normal code flow
because the Ruby code will explicitly expire the lock itself unless it
crashes mid way through runing.

The current lock ttl is designed to accomadate the max runtime the
cleanup process will run for before stopping. But since this is enforced
inside the worker execution path there is still room for the runtime to
exceed this limit. Since we only check that we've exceeded the runtime
limit after each query it's possible to have an additional 15 seconds on
the max runtime due to the last query taking 15 seconds (our Postgres
query timeout limit). Additionally there are a few steps more in the
code once we stop running the cleanup process so we could even run for a
few more seconds after that. In general it can't hurt having a longer
`ttl` than we need because under normal operation the `ttl` won't ever
be used. It is a problem if the `ttl` is too short because it will get
expired while the worker is still running. As such we just add a 20s
buffer. In theory we could even make this buffer longer because it will
only be invoked if the worker is terminated abnormally but then the
larger we make this ttl the longer we wait before the processing starts
again after the worker is terminated.