Fix sidekiq-cron poll interval to scale with number of process
On https://gitlab.com/gitlab-org/gitlab/-/merge_requests/99030 we
purported to allow sidekiq-cron to base its polling interval on
`scaled_poll_interval`. In that MR we gave precedence to
`Sidekiq.options[:cron_poll_interval]` but we neglected to note that it
is always set to a default (30s), which meant `scaled_poll_interval` was
never used.

Here we switch to instead check for
Gitlab.config.cron_jobs.poll_interval, which we know stays `nil` unless
overwritten by the user.