Fix flaky tests in spec/features/groups/issues_spec.rb
The tests in `spec/features/groups/issues_spec.rb` attempt to load a
list of issues and then drag them to a different order. However, if
the dragging attempts to start before they're actually loaded, the
JavaScript can fail with an error. To make this test more reliable,
wait for the requests to be done before attempting to execute
JavaScript.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/378591