Add committer name check to push_rules
Add a column named `commit_committer_name_check` to push_rules. If this option is enabled, we will check whether the git user name of the user is consistent with the gitlab user name when the user pushes the code. If not, the push will be rejected.

Changelog: added
EE: true