Use MessagePack for DuplicateJobs cookies
This updates the implementation of DuplicateJobs cookies (which is off
by default) to use MessagePack cookies. This prevents cookies from
growing large in case of a high deduplication factor.

While we are here, we also fix a race condition in the case a cookie
expires between SET NX and GET.

Changelog: other