Support structured error for UserCommitFiles
Gitaly is in the process of converting to structured errors for the
`UserCommitFiles` RPC, which allow more detailed errors to be returned
to Rails. This will eventually replace the error embedded in
`UserCommitFilesResponse`.

Update `#user_commit_files` to handle both the old error type and the
new structured errors.

Changelog: changed