Ensure auto-build-image and auto-deploy-image exist
auto-build-image and auto-deploy-image are automatically released on git
tags, but sometimes the release pipeline fails. Renovate watches git
tags when creating update MRs, so it's possible for it to create MRs for
images that have yet to be published.

This highlighted low hanging fruit for a smoke test: Ensure the image in
the template exists.