Allow setting max PG replicas rails process connects to
This adds a new field to our service discovery configuration in
`config/database.yml`. This field `max_replica_pools` controls how many
Postgres replicas that GitLab Rails connects.

GitLab.com runs multiple
replicas and each Postgres replica has multiple PGBouncer proxies in
front of it. These PGBouncer proxies are on the same host (hence same
IP address) but listen on different ports. Prior to this change all
GitLab Rails deployments would connect to all of these PGBouncer ports
but this is quite wasteful as effectively this means each Rails
deployment has many connections to the same database. Now that we're
approaching the limit of connections that 1 PGBouncer deployment can
reasonably handle we need a way to reduce the number of connections to
each PGBouncer deployment. This MR makes it possible by setting a
configurable limit.

The MR also theoretically can limit the total number
of replica databases (Postgres, not PGBouncer) that we connect to but as
long as the limit setting is greater than or equal to the number of
Postgres replicas we run then it will always preferentially choose 1
PGBouncer proxy for each Postgres server. That is the reason the Sampler
logic is quite complicated and looping over the `hostname` choosing 1
address at a time.

The other complex problem this needs to solve is the random sampling.
Service discovery runs every 1 minute on each GitLab Rails process.
Every minute we check for new replicas and when we detect the list of
replicas has changed we close all the connections and open new
connections to the replicas. Since this is expensive we don't want our
random sampler to choose a new set of random replicas every time it
runs. As such we make our random sampler choose exactly the same set of
hosts every time it runs by memoizing the `seed` at the time of
initialization of the `Sampler` class which should only be initialized
once when GitLab starts up.

Read more about this issue at
https://gitlab.com/gitlab-org/gitlab/-/issues/368625

Changelog: added