Remove ambiguous reference specs
These specs are causing gitaly to return GRPC internal errors. This is
equivalent to a 500 error in http. We should not be relying on such
behaviour.

While removing these tests will cause a temporary loss of
coverage, they have to be removed in order to release a fix which
handles this case properly.