Put tests in a describe block to avoid flaky tests
It is recommended to use `let_it_be` for the `create(:project)` setup.
However, this causes an issue because the access level setting changes
are persisted on to other tests. To prevent that from happening,
the tests using `let_it_be(:project) { create(:project) }` are put
in their own describe blocks