Handle password manager auto-fill-then-submit
Some users use password manager auto-fill-then-submit feature. When the
ArkoseLabs challenge is suppressed (not shown), this causes the login
form to be submitted before a session token from ArkoseLabs is received
which results to a failed login.

To handle this scenario, we wait for all calls to checkIfNeedsChallenge
to finish before deciding that a user does not need to be shown the
challenge and submit the form.

Changelog: fixed
EE: true