Remove unused UsersSelect QA helper
In the context of https://gitlab.com/gitlab-org/gitlab/-/merge_requests/101383
I have found that the Page::Component::UsersSelect helper seems to be
unused, as we have usage of `.ajax-users-dropdown` in the Frontend code
base any more and no QA specs relating to it are failing.

So let's remove the helper and it's only import.