Add expand syntax to CI config
This is the first step of implementing the raw variable support to
CI config.

The changes are behind the FF ci_raw_variables_in_yaml_config.

In `YamlProcessor::Result#root_variables` and `Entry::Variable`, I used
`YamlProcessor::FeatureFlags.enabled?`. So, I could only call these
within the scope of `YamlProcessor::FeatureFlags.with_actor`. However,
we call `root_variables` outside of the scope, so in this commit,
I moved `root_variables` into a getter method and now it is calculated
with the object instance.