Correctly position tree view with performance bar
When the performance bar is enabled it eats up vertical space at the
top. That's why the performance bar height is subtracted from the
max-height of the tree-view. But the top position does not take the
performance bar height into account.

This change adds the height of the, optional visible, performance bar to
the top position of the merge request tree view.

Changelog: fixed