Allow overriding GITALY_SERVER_VERSION in the env
The version of Gitaly is specified in the GITALY_SERVER_VERSION file in
the root of this project. For testing purposes it might be useful if
this can be easily overridden.

This change will read $GITALY_SERVER_VERSION from the environment, and
if it's set it ignores the file but uses the version specified in the
environment variable.