Add new fonts and fonts definitions
This adds GitLab Sans and JetBrains Mono as the new default fonts,
behind a feature flag. The new fonts will be delivered if the
`new_fonts` feature flag is enabled _or_ the `new_fonts` query parameter
is set.

The fonts are added to the Rails asset pipeline from an npm package
called `@gitlab/fonts`. This means we do not need to check in font files
in the GitLab repo.

Right now the feature flag will not work until:
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/3014 is merged
and updated inside GitLab.

Co-authored-by: Lukas Eipert <leipert@gitlab.com>
Co-authored-by: Thomas Hutterer <thutterer@gitlab.com>