Stop requesting gitaly disk access in repository check specs
It will not be possible to access gitaly controlled directories with
gitaly cluster as it will not be possible to locate the specific
repositories. Specs that need to manipulate git repositories directly
need to be moved to gitaly.

Since we cannot make invalid repositories anymore, instead we rely on
stubbing.