Remove unused blob_edit_content component
Apparently this component was introduced in 8b05626ba969c7da42695901e84
but it hasn't been utilized since 774cee04a4272a921ef1e8544bb2b0d01122.
This hasn't been a problem so far, but with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/96759
we added more utils to the `~/blob/utils.js`. Adding those utils means
that on every page now the `initSourceEditor` (and therefore Monaco!)
was loaded as well. Let's just remove the unused components and we
should be seeing improvement in our LCP numbers.