Potential fix for flaky compliance dashboard spec
This spec might be flaky because the first merge request defaults with
a `nil` `merged_at` value, but the test runs in this sequence:

```
visit group_security_compliance_dashboard_path(group)
merge_request.metrics.update!(merged_at: merged_at)
wait_for_requests
```

If the page is loaded before the metrics are updated, it's possible
that a row is missing from the dashboard.

To fix this, visit the page only when the merge request metrics have
been updated.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/381137